feat: added nickname to UserDetailsDialog
and made some refactoring to share the components between UserDetails
and ConversationSettings
